-initial commit Test task - Tale game
Description:
The page consists of a board of tiles (default 4x4).
- every 2 tiles have the same color (it could be an image, it doesn't matter);
- each tile is closed, so the user can not see a color (image).
The gameplay is the sequence of rounds. In each round the user should select 2 tiles with the same color to make them disappear. If he selects 2 tiles with different colors then they are flipped to "closed" state, and the user proceeds with the next round. The game is considered to be over when all of the tiles are opened.